Move subprocess tests to separate module
