+++ tests/test_commandline.py
@@ -4,7 +4,6 @@ import sys
 import tempfile
 from io import StringIO, BytesIO
 import pytest
-import subprocess
 
 from cutadapt.__main__ import main
 from utils import assert_files_equal, datapath, cutpath
@@ -610,10 +609,6 @@ def test_negative_length(run):
     run('--length -5', 'shortened-negative.fastq', 'small.fastq')
 
 
-def test_run_cutadapt_process():
-    subprocess.check_call(['cutadapt', '--version'])
-
-
 @pytest.mark.timeout(0.5)
 def test_issue_296(tmpdir):
     # Hang when using both --no-trim and --info-file together
@@ -661,68 +656,6 @@ def test_paired_separate(run):
     """test separate trimming of paired-end reads"""
     run("-a TTAGACATAT", "paired-separate.1.fastq", "paired.1.fastq")
     run("-a CAGTGGAGTA", "paired-separate.2.fastq", "paired.2.fastq")
-
-
-def test_run_as_module():
-    """Check that "python3 -m cutadapt ..." works"""
-    from cutadapt import __version__
-    py = subprocess.Popen([sys.executable, "-m", "cutadapt", "--version"], stdout=subprocess.PIPE)
-    assert py.communicate()[0].decode().strip() == __version__
-
-
-def test_standard_input_pipe(tmpdir, cores):
-    """Read FASTQ from standard input"""
-    out_path = str(tmpdir.join("out.fastq"))
-    in_path = datapath("small.fastq")
-    # Use 'cat' to simulate that no file name is available for stdin
-    with subprocess.Popen(["cat", in_path], stdout=subprocess.PIPE) as cat:
-        with subprocess.Popen([
-            sys.executable, "-m", "cutadapt", "--cores", str(cores),
-            "-a", "TTAGACATATCTCCGTCG", "-o", out_path, "-"],
-            stdin=cat.stdout
-        ) as py:
-            _ = py.communicate()
-            cat.stdout.close()
-            _ = py.communicate()[0]
-    assert_files_equal(cutpath("small.fastq"), out_path)
-
-
-def test_standard_output(tmpdir, cores):
-    """Write FASTQ to standard output (not using --output/-o option)"""
-    out_path = str(tmpdir.join("out.fastq"))
-    with open(out_path, "w") as out_file:
-        py = subprocess.Popen([
-            sys.executable, "-m", "cutadapt", "--cores", str(cores),
-            "-a", "TTAGACATATCTCCGTCG", datapath("small.fastq")],
-            stdout=out_file)
-        _ = py.communicate()
-    assert_files_equal(cutpath("small.fastq"), out_path)
-
-
-def test_explicit_standard_output(tmpdir, cores):
-    """Write FASTQ to standard output (using "-o -")"""
-
-    out_path = str(tmpdir.join("out.fastq"))
-    with open(out_path, "w") as out_file:
-        py = subprocess.Popen([
-            sys.executable, "-m", "cutadapt", "-o", "-", "--cores", str(cores),
-            "-a", "TTAGACATATCTCCGTCG", datapath("small.fastq")],
-            stdout=out_file)
-        _ = py.communicate()
-    assert_files_equal(cutpath("small.fastq"), out_path)
-
-
-def test_force_fasta_output(tmpdir, cores):
-    """Write FASTA to standard output even on FASTQ input"""
-
-    out_path = str(tmpdir.join("out.fasta"))
-    with open(out_path, "w") as out_file:
-        py = subprocess.Popen([
-            sys.executable, "-m", "cutadapt", "--fasta", "-o", "-", "--cores", str(cores),
-            "-a", "TTAGACATATCTCCGTCG", datapath("small.fastq")],
-            stdout=out_file)
-        _ = py.communicate()
-    assert_files_equal(cutpath("small.fasta"), out_path)
 
 
 def test_empty_read_with_wildcard_in_adapter(run):
